voice_DZF/requirements.txt
SpeechRecognition
sounddevice
soundfile
scipy
psycopg2-binary
python-dotenv
